refactor: relative-time helpers move to a shared private module
_relative_time_node and _is_clock_read (plus the _TIME_FUNCS and
_TIME_FUNC_NAMES collections they depend on) move from
validation/examples.py to a new validation/_timewindow.py, mirroring the
existing _scalar.py split. examples.py imports Validator from
validator.py, so validator.py (needed by an upcoming task) could not
import the helper back from examples.py without a circular import.

examples.py re-imports _relative_time_node (its remaining caller) plus
_TIME_FUNCS/_TIME_FUNC_NAMES (re-exported, aliased to themselves, since
tests/test_validation/test_examples.py imports them from that module
directly). _is_clock_read has no caller outside _relative_time_node, so
it is not re-imported.

Updates the sqlglot>=28.6 floor comment in pyproject.toml to point at
the new module.
